pyproject.toml
[tool.poetry]
name = "imagetextextractor"
version = "0.1.0"
description = "An OCR application that extracts text from images."
authors = ["Arkeodev <arkeodev@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
pillow = "^10.4.0"
streamlit = "^1.40.1"
langchain = "^0.3.8"
langchain-community = "^0.3.8"
openai = "^1.55.0"
together = "1.3.5"
langchain-core = "^0.3.21"
langchain-together = "^0.2.0"
python-multipart = "^0.0.17"
requests = "^2.32.3"

[tool.poetry.dev-dependencies]
pytest = "^8.3.3"
pytest-mock = "^3.14.0"
isort = "^5.13.2"
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
